% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{addTextToData}
\alias{addTextToData}
\title{addTextToData}
\usage{
addTextToData(df, lang, topic)
}
\arguments{
\item{df}{dataframe}

\item{lang}{language}

\item{topic}{topic}
}
\value{
dataframe

examples see function return_data
}
\description{
Adds variables containing text labels to the id variables.
}
