% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{contentToDataframe_data}
\alias{contentToDataframe_data}
\title{contentToDataframe_data}
\usage{
contentToDataframe_data(resp_object_list, addText)
}
\arguments{
\item{resp_object_list}{list of S3 objects of class hwdb_api}

\item{addText}{logical indicating whether variables containing text labels
should be added to the id variables, the former variables
will be returned as ordered factors, except for diagnosis and
year which is left to the user to handle. The ordering follows the
value of the id variable.}
}
\value{
data frame

examples see function return_data
}
\description{
Converts content of list of S3 objects of class hwdb_api  to data frame
}
