% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{contentToDataframe_meta}
\alias{contentToDataframe_meta}
\title{contentToDataframe_meta}
\usage{
contentToDataframe_meta(resp_object)
}
\arguments{
\item{resp_object}{S3 object of class hwdb_api}
}
\value{
dataframe

examples: see function return_meta
}
\description{
Converts content of meta data S3 objects of class hwdb_api  to data frame
}
