% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{get_apiDoc_url}
\alias{get_apiDoc_url}
\title{get_apiDoc_url}
\usage{
get_apiDoc_url()
}
\value{
string with url to API documentation

examples
get_apiDoc_url()
}
\description{
Return path of documentation of API
}
