% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{get_Rpkg_apiVersion}
\alias{get_Rpkg_apiVersion}
\title{get_Rpkg_apiVersion}
\usage{
get_Rpkg_apiVersion()
}
\value{
string with version of API for which the package is
developed

examples
get_Rpkg_apiVersion()
}
\description{
Returns version of the API used in this R package
}
