% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{hwdb_api}
\alias{hwdb_api}
\title{hwdb_api}
\usage{
hwdb_api(version, lang, topic, resultquery)
}
\arguments{
\item{version}{\itemize{
\item API version
}}

\item{lang}{\itemize{
\item language
}}

\item{topic}{\itemize{
\item topic that should be queried
}}

\item{resultquery}{\itemize{
\item part of the query that is the remainder after topic
}}
}
\value{
server response S3 object of class "hwdb_api"

examples
}
\description{
Sends request to server
}
