% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{print.hwdb_api}
\alias{print.hwdb_api}
\title{print}
\usage{
\method{print}{hwdb_api}(x, ...)
}
\arguments{
\item{x}{object of type hwdb_api}

\item{...}{arguments to be passed to methods

examples}
}
\description{
Print function for hwdb_api object
}
