% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{return_data}
\alias{return_data}
\title{return_data}
\usage{
return_data(lang, topic, df_input_vars, addText)
}
\arguments{
\item{lang}{language}

\item{topic}{topic of data}

\item{df_input_vars}{data frame with two variables: var_list and values_list,
for each of the variables available for the topic it
should contain the values for which data is desired, see
example below}

\item{addText}{logical indicating whether variables containing text labels
should be added to the id variables, the former variables
will be returned as ordered factors}
}
\value{
XXX
}
\description{
Returns data from query
}
