% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{return_meta}
\alias{return_meta}
\title{return_meta}
\usage{
return_meta(type, lang, topic, var)
}
\arguments{
\item{type}{type of metadata, should be one of:
"api_version" for version of api
"lang" for language
"topic" of data
"var" for the variable}

\item{lang}{language (depends on version)}

\item{topic}{topic of data (depends on language)}

\item{var}{variable (depends on topic)}
}
\value{
dataframe with id and text variables, sometimes also additional
variables
}
\description{
Returns API metadata, lists all values for the queried parameter
}
\examples{

versions <- return_meta(type="api_version")
print(versions)

languages <- return_meta(type="lang")
print(languages)

topics <- return_meta(type="topic", lang="en")
print(topics)
vars <- return_meta(type="var", lang="en", topic="diagnoserislutenvard")
print(vars[,1:2])
var_cats_alder <- return_meta(type="var_cat", lang="en", 
                   topic="diagnoserislutenvard", var="alder")
print(var_cats_alder)
}
