% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{update_url}
\alias{update_url}
\title{update_url
Modify url based on the additional path of the query}
\usage{
update_url(base, path)
}
\arguments{
\item{base}{base url to the API}

\item{path}{API url path}
}
\value{
string with the complete url

examples
}
\description{
update_url
Modify url based on the additional path of the query
}
