% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{return_data}
\alias{return_data}
\title{return_data}
\usage{
return_data(lang, topic, df_input_vars, addText)
}
\arguments{
\item{lang}{language}

\item{topic}{topic of data}

\item{df_input_vars}{data frame with two variables: var_list and values_list,
for each of the variables available for the topic it
should contain the values for which data is desired, see
example below}

\item{addText}{logical indicating whether variables containing text labels
should be added to the id variables, the former variables
will be returned as ordered factors except for diagnosis and year
which is left to the user to handle. The ordering follows the
value of the id variable.}
}
\value{
data frame
}
\description{
Returns data from query
}
\examples{
vars <- return_meta(type="var", lang="en", topic="diagnoserislutenvard")
var_list <- vars[,1]
#- Region=entire Sweden, age group=45-49, sex= males and females, measure=
# Number of patients per 100,000 inhabitants years = 2012 and 2013, 
# diagnoses= J13 and J14
values_list <- c("0",
                 "10",
                 "1,2",
                  "7",
                  "2012,2013",
                  "J13,J14")

df_input_vars <- as.data.frame(cbind(var_list,values_list))
print(df_input_vars)

data <- return_data(lang="en",topic="diagnoserislutenvard", df_input_vars, 
addText=TRUE)

}
