% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoS_functions.R
\name{return_meta}
\alias{return_meta}
\title{return_meta}
\usage{
return_meta(type, lang, topic, var)
}
\arguments{
\item{type}{type of metadata, should be one of:
"api_version" for version of api
"lang" for language
"topic" of data
"var" for the variable}

\item{lang}{language (depends on version)}

\item{topic}{topic of data (depends on language)}

\item{var}{variable (depends on topic)}
}
\value{
S3 object of type "hwdb_api"
}
\description{
Returns API metadata, lists all values
}
